chore: 🔧 remove pointer capture events
