+++ src/components/availability/table/availability-block-cell.tsx
@@ -1,11 +1,7 @@
-import { useCallback, useRef } from "react";
 import { useShallow } from "zustand/shallow";
 import { AvailabilityBlock } from "@/components/availability/table/availability-block";
 import { GoogleCalendarEventBlock } from "@/components/availability/table/google-calendar-event-block";
-import type {
-	AvailabilityBlockType,
-	EventSegment,
-} from "@/lib/types/availability";
+import type { EventSegment } from "@/lib/types/availability";
 import { cn } from "@/lib/utils";
 import { useBlockSelectionStore } from "@/store/useBlockSelectionStore";
 
@@ -13,7 +9,6 @@ interface AvailabilityBlockCellProps {
 	blockIndex: number;
 	isAvailable: boolean;
 	zotDateIndex: number;
-	setAvailabilities: (startBlock: AvailabilityBlockType) => void;
 	isTopOfHour: boolean;
 	isHalfHour: boolean;
 	isLastRow: boolean;
@@ -25,96 +20,22 @@ export function AvailabilityBlockCell({
 	blockIndex,
 	isAvailable,
 	zotDateIndex,
-	setAvailabilities,
 	isTopOfHour,
 	isHalfHour,
 	isLastRow,
 	eventSegments,
 	hasSpacerBefore = false,
 }: AvailabilityBlockCellProps) {
-	const isDraggingRef = useRef(false);
-
-	const {
-		startBlockSelection,
-		selectionState,
-		setStartBlockSelection,
-		setEndBlockSelection,
-	} = useBlockSelectionStore(
+	const { selectionState } = useBlockSelectionStore(
 		useShallow((state) => ({
-			startBlockSelection: state.startBlockSelection,
 			selectionState: state.selectionState,
-			setStartBlockSelection: state.setStartBlockSelection,
-			setEndBlockSelection: state.setEndBlockSelection,
 		})),
-	);
-
-	// Pointer event handlers
-	const handlePointerDown = useCallback(
-		(e: React.PointerEvent) => {
-			const target = e.currentTarget as HTMLElement;
-			target.setPointerCapture(e.pointerId);
-
-			isDraggingRef.current = true;
-			const selection = { zotDateIndex, blockIndex };
-			setStartBlockSelection(selection);
-			setEndBlockSelection(selection);
-		},
-		[zotDateIndex, blockIndex, setStartBlockSelection, setEndBlockSelection],
-	);
-
-	const handlePointerMove = useCallback(
-		(e: React.PointerEvent) => {
-			if (!isDraggingRef.current) return;
-
-			const element = document.elementFromPoint(e.clientX, e.clientY);
-			if (!element) return;
-
-			const touchingDateIndex = parseInt(
-				element.getAttribute("data-date-index") || "",
-				10,
-			);
-			const touchingBlockIndex = parseInt(
-				element.getAttribute("data-block-index") || "",
-				10,
-			);
-
-			if (
-				!Number.isNaN(touchingDateIndex) &&
-				!Number.isNaN(touchingBlockIndex)
-			) {
-				setEndBlockSelection({
-					zotDateIndex: touchingDateIndex,
-					blockIndex: touchingBlockIndex,
-				});
-			}
-		},
-		[setEndBlockSelection],
-	);
-
-	const handlePointerUp = useCallback(
-		(e: React.PointerEvent) => {
-			if (!isDraggingRef.current) return;
-
-			const target = e.currentTarget as HTMLElement;
-			if (target.hasPointerCapture(e.pointerId)) {
-				target.releasePointerCapture(e.pointerId);
-			}
-
-			isDraggingRef.current = false;
-			if (startBlockSelection) {
-				setAvailabilities(startBlockSelection);
-			}
-		},
-		[startBlockSelection, setAvailabilities],
 	);
 
 	return (
 		<td className="relative px-0 py-0">
 			<button
 				type="button"
-				onPointerDown={handlePointerDown}
-				onPointerMove={handlePointerMove}
-				onPointerUp={handlePointerUp}
 				data-date-index={zotDateIndex}
 				data-block-index={blockIndex}
 				className={cn(
